fix(login): image env
